Use Vundle instead of Pathogen. Automate setup
All plugins in .vimrc are now automatically installed/updated when init
script is run. Vundle documentation is available at:
https://github.com/gmarik/Vundle.vim
Also, a registry of vim plugins is availble at: http://vimawesome.com/

New in the generic automation: any scripts in the folder `setup_scripts`
will be executed. If they are not executable or exit with a non-zero
exit status, the init script will fail.