Work on a Mac
Mac's have older / different tools mostly. Have to install Homebrew and
get the GNU things we need before the init script can actually be run.

Change all shebangs that use /bin/bash to /usr/bin/env bash because the
bash that is installed by brew lives elsewhere. And that is the one I
want to use.