Switch back to guake from tilda
Tilda was being a bit of a pain. The original reason for switching away
from guake was no way to configure as code.

Discovered a workaround:
https://github.com/Guake/guake/issues/292#issuecomment-408930278

Against tilda:
- Using tilda required hacking the desktop file to point to the config
we want.
- Tilda had to be closed for any config changes made directly to the
config file, otherwise would be overwritten when tilda closed.
- Tilda didn't have an easy way to switch between a light and dark theme
which is important for working at night vs working outside at a cafe.